Add 'req'd userdata variables' directive and more tests
